requirements.txt
fastapi==0.115.*
pydantic==2.11.*
uvicorn[standard]==0.31.*
python-dotenv==1.*

# settings
pydantic-settings==2.5.2

# HTTP 클라이언트(비동기)
httpx==0.27.*

# (있으면 그대로 두기)
requests==2.32.*

# RAG / 임베딩 관련
langchain==0.1.*
langchain-community==0.0.*
sentence-transformers==2.2.*
torch==2.6.*
faiss-cpu==1.9.*
pypdf==3.17.*
sqlalchemy==2.0.*
pgvector==0.2.*

# MCP
fastmcp
mcp>=1.22.0
pg8000